[CIDMDEV-5136] - Migration Support - Read-Only Region
- Protected the following services by flag 'feature.migration.read.only.mode.enabled'
(returns 503 when read-only enabled) :

1. POST v2.0/OS-KSADM/services
2. DELETE v2.0/RAX-AUTH/regions/{name}
3. PUT v2.0/RAX-AUTH/regions/{name}
4. PUT v2.0/RAX-AUTH/default-region/services

The last service (PUT RAX-AUTH/default-region/services) updates the application to set the default region services. The setting doesn't appear to be used at all. While this updates the application (rather than a region), it was protected as part of this story.

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -151,7 +151,11 @@ public class Cloud20VersionResource {
     @Path("RAX-AUTH/default-region/services")
     public Response setDefaultRegionServices(@HeaderParam(X_AUTH_TOKEN) String authToken,
                                              DefaultRegionServices defaultRegionServices){
-        return cloud20Service.setDefaultRegionServices(authToken, defaultRegionServices).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.setDefaultRegionServices(authToken, defaultRegionServices).build();
+        }
     }
 
     @POST
@@ -1094,7 +1098,11 @@ public class Cloud20VersionResource {
     @POST
     @Path("RAX-AUTH/regions")
     public Response createRegion(@Context UriInfo uriInfo, @HeaderParam(X_AUTH_TOKEN) String authToken, Region region) {
-        return cloud20Service.addRegion(uriInfo, authToken, region).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addRegion(uriInfo, authToken, region).build();
+        }
     }
 
     @GET
@@ -1112,13 +1120,21 @@ public class Cloud20VersionResource {
     @PUT
     @Path("RAX-AUTH/regions/{name}")
     public Response updateRegion(@HeaderParam(X_AUTH_TOKEN) String authToken, @PathParam("name") String name, Region region) {
-        return cloud20Service.updateRegion(authToken, name, region).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.updateRegion(authToken, name, region).build();
+        }
     }
 
     @DELETE
     @Path("RAX-AUTH/regions/{name}")
     public Response deleteRegion(@HeaderParam(X_AUTH_TOKEN) String authToken, @PathParam("name") String name) {
-        return cloud20Service.deleteRegion(authToken, name).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deleteRegion(authToken, name).build();
+        }
     }
 
     @POST
@@ -1212,6 +1228,10 @@ public class Cloud20VersionResource {
 
     private boolean isSamlEnabled(){
        return config.getBoolean("saml.enabled");
+    }
+
+    public void setIdentityConfig(IdentityConfig identityConfig) {
+        this.identityConfig = identityConfig;
     }
 }
